refactor to manager
